refactor(conformance): split runner helper utilities
